More routes to Flask API, better abort
- /init initialized DB and uploads geo data
- /analyze?size=<number> run analysis
